Add files via upload
There are two classes called NumerTest and StringTest. Each Class has two method. There are "addNumber" and "subNumber" in NumberTest, "concatString" and "equalString" in StringTest.
When a user says the name of a class and then the name of a method, that method is called and executed Otherwise an error message will be printed.